working on event form
Can now create an event and it will add to the database. You can also view the details of this event.
-added for loops for viewing events in the database for event.html and index.html

Event.html: needs to be filtered by logged in user - can't figure it out

Index.html: also isnt showing events im not sure why.

Hopefully having written the for loops will make it easier for you guys to adjust it slightly.

Also image upload now works (the file path was missing an "s" in the function)